✨<Presentation> 사진을 추가하면 slider 화면으로 넘어간다.
